♻️ Move business logic to core package
