+++ core/options.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	"github.com/akm/git-exec/git"
