+++ main.go
@@ -4,20 +4,22 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+
+	"github.com/akm/git-exec/core"
 )
 
 func main() {
 	if len(os.Args) < 2 {
-		Help()
+		core.Help()
 		os.Exit(1)
 	}
 
-	options, commandArgs, err := ParseOptions(os.Args[1:])
+	options, commandArgs, err := core.ParseOptions(os.Args[1:])
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Failed to parse arguments: %s\n", err.Error())
 	}
 	if options.Help {
-		Help()
+		core.Help()
 		os.Exit(0)
 	} else if options.Version {
 		if len(commandArgs) == 0 {
@@ -28,7 +30,7 @@ func main() {
 		}
 	}
 
-	if err := Run(options, commandArgs); err != nil {
+	if err := core.Run(options, commandArgs); err != nil {
 		fmt.Fprintln(os.Stderr, err.Error())
 		os.Exit(1)
 	}
